# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LeaderboardEntry {
  rank: number;
  uuid: string;
  username: string;
  region?: string | null;
  mmr: number;
  points: number;
  wins: number;
  losses: number;
  fights: number;
  tier: string;
  isHT1: boolean;
  progress: number;
}

export interface LeaderboardResponse {
  mode: string;
  entries: LeaderboardEntry[];
}

export interface OverviewEntry {
  rank: number;
  uuid: string;
  username: string;
  region?: string | null;
  totalMMR: number;
  totalPoints: number;
  rankedModes: number;
  totalWins: number;
  totalLosses: number;
}

export interface OverviewLeaderboardResponse {
  entries: OverviewEntry[];
}

export interface ModeStats {
  mmr: number;
  wins: number;
  losses: number;
  fights: number;
  tier: string;
  isHT1: boolean;
  progress: number;
  placed: boolean;
}

export type PlayerProfileResponseModes = { [key: string]: ModeStats };

export interface PlayerProfileResponse {
  uuid: string;
  username: string;
  region?: string | null;
  modes: PlayerProfileResponseModes;
  totalMMR: number;
  totalWins: number;
  totalLosses: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetOverviewLeaderboardParams = {
  limit?: number;
};